chore: consolidate aws to a single file
